requirements.txt
pynrrd
numpy
scipy
matplotlib
scikit-image
patato
pacfish
simpa